move extension to main app
