+++ src/apps/main/nautilus-extension/install.ts
