+++ src/apps/main/nautilus-extension/service.ts
@@ -2,7 +2,7 @@ import Logger from 'electron-log';
 import fs from 'fs/promises';
 import os from 'os';
 import path from 'path';
-import { doesFileExist } from '../shared/fs/fileExists';
+import { doesFileExist } from '../../shared/fs/fileExists';
 import { exec } from 'child_process';
 
 const name = 'internxt-virtual-drive.py';
