+++ src/apps/main/nautilus-extension/uninstall.ts
